+++ Source/DataTransfer.Legacy.MassImport.NUnit.Integration/FunctionalTests/CollationEdgeCasesTests.cs
@@ -5,7 +5,7 @@ using MassImport.NUnit.Integration.Helpers;
 using NUnit.Framework;
 using Relativity;
 using Relativity.Core.Service;
-using Relativity.MassImport;
+using Relativity.MassImport.DTO;
 using Relativity.MassImport.Api;
 using MassImportManager = Relativity.MassImport.Api.MassImportManager;