+++ Source/DataTransfer.Legacy.MassImport/Data/ChoicesImportSql.cs
@@ -10,9 +10,9 @@ namespace Relativity.MassImport.Data
 		public TableNames TableNames { get; private set; }
 		public string KeyColumnName { get; private set; }
 
-		public ChoicesImportSql(TableNames tableNames, string keyColumnName, OverwriteType overwriteType)
+		public ChoicesImportSql(TableNames tableNames, string keyColumnName, Relativity.MassImport.DTO.OverwriteType overwriteType)
 		{
-			_isAppendOnly = overwriteType == OverwriteType.Append;
+			_isAppendOnly = overwriteType == Relativity.MassImport.DTO.OverwriteType.Append;
 			TableNames = tableNames;
 			KeyColumnName = keyColumnName;
 		}
@@ -91,7 +91,7 @@ WHERE
 		WHERE
 			ArtifactID = Z.[AssociatedArtifactID]
 			AND
-			N.[kCura_Import_Status] = {(long)ImportStatus.Pending}
+			N.[kCura_Import_Status] = {(long)Relativity.MassImport.DTO.ImportStatus.Pending}
 	);";
 		}
 
@@ -111,7 +111,7 @@ AS
 		N.[{KeyColumnName}] = C.[DocumentIdentifier]
 	WHERE
 		C.[CodeTypeID] = {codeTypeId}
-		AND N.[kCura_Import_Status] = {(long)ImportStatus.Pending}
+		AND N.[kCura_Import_Status] = {(long)Relativity.MassImport.DTO.ImportStatus.Pending}
 )
 
 INSERT INTO {codeMappingTable} ([CodeArtifactID], [AssociatedArtifactID])