+++ Source/DataTransfer.Legacy.MassImport/Data/Helper.cs
@@ -150,19 +150,19 @@ FROM
 			string result;
 			switch (settings.OverlayBehavior)
 			{
-				case OverlayBehavior.MergeAll:
+				case Relativity.MassImport.DTO.OverlayBehavior.MergeAll:
 					{
 						result = "1";
 						break;
 					}
 
-				case OverlayBehavior.ReplaceAll:
+				case Relativity.MassImport.DTO.OverlayBehavior.ReplaceAll:
 					{
 						result = "0";
 						break;
 					}
 
-				case OverlayBehavior.UseRelativityDefaults:
+				case Relativity.MassImport.DTO.OverlayBehavior.UseRelativityDefaults:
 					{
 						if (overlayMergeValues.HasValue && overlayMergeValues.Value)
 						{
@@ -185,7 +185,7 @@ FROM
 			return result;
 		}
 
-		public static bool IsMergeOverlayBehavior(OverlayBehavior overlayBehavior, FieldTypeHelper.FieldType fieldType, bool? overlayMergeValues)
+		public static bool IsMergeOverlayBehavior(Relativity.MassImport.DTO.OverlayBehavior overlayBehavior, FieldTypeHelper.FieldType fieldType, bool? overlayMergeValues)
 		{
 			// This function is called only for multi choices/objects
 			if (fieldType == FieldTypeHelper.FieldType.Code || fieldType == FieldTypeHelper.FieldType.Object)
@@ -195,17 +195,17 @@ FROM
 
 			switch (overlayBehavior)
 			{
-				case OverlayBehavior.MergeAll:
+				case Relativity.MassImport.DTO.OverlayBehavior.MergeAll:
 					{
 						return true;
 					}
 
-				case OverlayBehavior.ReplaceAll:
+				case Relativity.MassImport.DTO.OverlayBehavior.ReplaceAll:
 					{
 						return false;
 					}
 
-				case OverlayBehavior.UseRelativityDefaults:
+				case Relativity.MassImport.DTO.OverlayBehavior.UseRelativityDefaults:
 					{
 						return overlayMergeValues.HasValue && overlayMergeValues.Value;
 					}
@@ -232,7 +232,7 @@ FROM
 					using (var errorFile = new System.IO.StreamWriter(System.IO.Path.Combine(defaultLocation, errorFileName)))
 					{
 						while (reader.Read())
-							errorFile.WriteLine(string.Format("\"{1}{0}{2}{0}{3}\"", "\",\"", reader.GetInt32(0), ImportStatusHelper.GetCsvErrorLine(reader.GetInt64(1), reader.GetString(2), "", -1, reader.GetString(2), reader.IsDBNull(3) ? null : reader.GetString(3), reader.IsDBNull(4) ? null : reader.GetString(4)), reader.GetString(2)));
+							errorFile.WriteLine(string.Format("\"{1}{0}{2}{0}{3}\"", "\",\"", reader.GetInt32(0), Relativity.MassImport.DTO.ImportStatusHelper.GetCsvErrorLine(reader.GetInt64(1), reader.GetString(2), "", -1, reader.GetString(2), reader.IsDBNull(3) ? null : reader.GetString(3), reader.IsDBNull(4) ? null : reader.GetString(4)), reader.GetString(2)));
 					}
 				}
 			}
@@ -260,7 +260,7 @@ SELECT
 	[kCura_Import_ErrorData]
 FROM [Resource].[{ tableName }]
 WHERE
-	NOT [kCura_Import_Status] = { (long) ImportStatus.Pending }
+	NOT [kCura_Import_Status] = { (long)Relativity.MassImport.DTO.ImportStatus.Pending }
 ORDER BY
 	kCura_Import_OriginalLineNumber";